Fix for Issue #596
