+++ modules/jpa-envers/src/main/java/io/oasp/module/jpa/dataaccess/api/AdvancedRevisionEntity.java
@@ -6,6 +6,7 @@ import java.util.Date;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
 
@@ -19,12 +20,11 @@ import org.hibernate.envers.RevisionTimestamp;
  * This is a custom {@link org.hibernate.envers.DefaultRevisionEntity revision entity} also containing the actual user.
  *
  * @see org.hibernate.envers.DefaultRevisionEntity
- * @deprecated please use {@link io.oasp.module.jpa.dataaccess.api.common.AdvancedRevisionEntity} instead.
+ *
  */
 @Entity
 @RevisionEntity(AdvancedRevisionListener.class)
 @Table(name = "RevInfo")
-@Deprecated
 public class AdvancedRevisionEntity implements PersistenceEntity<Long> {
 
   /** UID for serialization. */
@@ -43,9 +43,9 @@ public class AdvancedRevisionEntity implements PersistenceEntity<Long> {
   /** @see #getDate() */
   private transient Date date;
 
-  /** @see #getUser() */
+  /** @see #getUserLogin() */
 
-  private String user;
+  private String userLogin;
 
   /**
    * The constructor.
@@ -100,20 +100,21 @@ public class AdvancedRevisionEntity implements PersistenceEntity<Long> {
    * @return the login or id of the user that has created this revision.
    */
 
-  public String getUser() {
+  public String getUserLogin() {
 
-    return this.user;
+    return this.userLogin;
   }
 
   /**
-   * @param user is the new value of {@link #getUser()}.
+   * @param userLogin is the new value of {@link #getUserLogin()}.
    */
-  public void setUser(String user) {
+  public void setUserLogin(String userLogin) {
 
-    this.user = user;
+    this.userLogin = userLogin;
   }
 
   @Override
+  @GeneratedValue(strategy = GenerationType.TABLE)
   public int getModificationCounter() {
 
     return 0;
