+++ modules/jpa-envers/src/main/java/io/oasp/module/jpa/dataaccess/api/AdvancedRevisionListener.java
@@ -6,8 +6,6 @@ import net.sf.mmm.util.session.api.UserSessionAccess;
 
 import org.hibernate.envers.RevisionListener;
 
-import io.oasp.module.jpa.dataaccess.api.common.AdvancedRevisionEntity;
-
 /**
  * This is the implementation of {@link RevisionListener} that enriches {@link AdvancedRevisionEntity} with additional
  * information.
